batch: Move ownership unit source builder
Batch ownership owns the unit model used by reset-style selection paths,
while display owns reconstructed batch-source views.

Source-line unit construction still rebuilds display lines from inside
ownership, coupling the ownership model back to display rendering.

This commit addresses that boundary by moving the source-line unit builder
into a new `batch.ownership_units` bridge and updating production callers.

The next commit will move unit-grouping tests to the bridge module.
