+++ src/git_stage_batch/batch/selection.py
@@ -9,11 +9,11 @@ from typing import TYPE_CHECKING, Optional
 
 from .ownership import (
     BatchOwnership,
-    build_ownership_units_from_batch_source_lines,
     select_ownership_units_by_display_ids,
     validate_ownership_units,
     rebuild_ownership_from_units,
 )
+from .ownership_units import build_ownership_units_from_batch_source_lines
 from ..core.line_selection import (
     LineRanges,
     LineSelection,
